requirements.txt
Flask
openai
python-docx
pdfminer.six
reportlab
pyinstaller
